\name{exdebug-package}
\alias{exdebug-package}
\alias{exdebug}
\docType{package}
\title{
\packageTitle{exdebug}
}
\description{
\packageDescription{exdebug}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{exdebug}
\packageIndices{exdebug}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{exdebug}

Maintainer: \packageMaintainer{exdebug}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
